+++ tests/QueryTest.php
@@ -1,0 +1,7 @@
+<?php
+    echo getcwd();
+    include_once(dirname(dirname(__FILE__)) . '/pop.php');
+
+    class QueryTest extends PHPUnit_Framework_TestCase {
+
+    }